Move caching repository to correct location
